+++ packages/extension/test/dag-view.test.tsx
@@ -39,6 +39,8 @@ const MODEL: ArchgenModelMessage = {
   codegraph: { product: 'unsupported' },
   themeKind: 'dark',
   warnings: [],
+  features: [],
+  activeSlug: '',
 };
 
 interface FakeApi {